requirements.txt
python-telegram-bot>=20,<22
python-dotenv>=1.0
pytz>=2024.1
SQLAlchemy>=2.0
google-api-python-client>=2.141.0
google-auth-oauthlib>=1.2.1
google-auth>=2.34.0
